fix: ensure grid resizing uses size_t for width and height in Maze constructor

+++ src/include/maze.hh
@@ -37,7 +37,7 @@ class Maze {
     std::vector<Cell> grid;
 
     Maze(int width, int height) : width(width), height(height) {
-        grid.resize(width * height);
+        grid.resize(static_cast<size_t>(width) * static_cast<size_t>(height));
         generate();
     }
